asset: limit max value of issuance leaf to math.MaxUint32
This is a step towards fixing:
https://github.com/lightninglabs/taproot-assets/issues/544.

In this commit, we limit the max amount of an asset that can be issued
at one time to math.MaxUint32. The existing 32-bit sum is left in place.
This gives us enough buffer room, as we'd need 4 billion UTXOs of the
same assets to overflow the root sum. In the future, we can lift this
limit with future asset versions.

An asset can have more than 4 billion ish units if they issue with
distinct UTXOs. This value can be viewed as the max amt that can be
issued in a single tranche.

+++ address/address_test.go
@@ -43,7 +43,7 @@ func randAddress(t *testing.T, net *ChainParams, v Version, groupPubKey,
 	}
 
 	if amt == nil && assetType == asset.Normal {
-		amount = test.RandInt[uint64]()
+		amount = uint64(test.RandInt[uint32]())
 	}
 
 	var tapscriptSibling *commitment.TapscriptPreimage
